change project to JSTL, delete AssignServlet

+++ src/main/java/servlets/DeleteServlet.java
@@ -13,19 +13,26 @@ import java.sql.SQLException;
 
 @WebServlet("/admin/delete")
 public class DeleteServlet extends HttpServlet {
-    UserService userService = UserService.getInstance();
+    private UserService userService = UserService.getInstance();
 
 
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String id = request.getParameter("allUsers");
 
-
-        try {
-            User user = userService.getUserById(Long.parseLong(id));
-            userService.deleteUser(user);
-        } catch (SQLException e) {
-            e.printStackTrace();
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        
+        String[] id = req.getParameterValues("del");
+        if(id !=null ) {
+            for (String str : id) {
+                try {
+                    User user = userService.getUserById(Long.parseLong(str));
+                    userService.deleteUser(user);
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            resp.sendRedirect("/admin");
+        }else {
+            resp.sendRedirect("/admin");
         }
-        response.sendRedirect("/admin");
     }
 }
